refactor(ic): the terminal detection moves to termgfx, and ic calls it
