+++ src/ic/Cargo.toml
@@ -18,6 +18,7 @@ image.workspace = true
 libc.workspace = true
 notify.workspace = true
 termion.workspace = true
+termgfx.workspace = true
 termsize.workspace = true
 unicode-width.workspace = true
